Develop 병합

+++ src/main/java/com/helpmeCookies/user/service/UserService.java
@@ -1,21 +1,17 @@
 package com.helpmeCookies.user.service;
 
-import java.io.IOException;
-import java.util.List;
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
+import com.helpmeCookies.global.exception.user.DuplicateResourceException;
 import com.helpmeCookies.global.exception.user.ResourceNotFoundException;
 import com.helpmeCookies.global.utils.AwsS3FileUtils;
 import com.helpmeCookies.user.dto.UserCommonInfoDto;
 import com.helpmeCookies.user.dto.UserDto;
 import com.helpmeCookies.user.dto.UserInfoDto;
 import com.helpmeCookies.user.dto.UserTypeDto;
-import com.helpmeCookies.user.dto.request.UserReq;
 import com.helpmeCookies.user.dto.response.UserFollowingRes;
 import com.helpmeCookies.user.entity.ArtistInfo;
 import com.helpmeCookies.user.entity.Social;
@@ -24,9 +20,10 @@ import com.helpmeCookies.user.entity.UserInfo;
 import com.helpmeCookies.user.repository.ArtistInfoRepository;
 import com.helpmeCookies.user.repository.SocialRepository;
 import com.helpmeCookies.user.repository.UserRepository;
-import com.sun.jdi.request.DuplicateRequestException;
 
 import lombok.RequiredArgsConstructor;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -39,8 +36,7 @@ public class UserService {
 
 	@Transactional
 	public UserDto getUserInfo(Long userId) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 유저입니다."));
+		User user = checkUser(userId);
 
 		return UserDto.fromEntity(user);
 	}
@@ -48,16 +44,17 @@ public class UserService {
 	@Transactional
 	public UserDto updateUser(UserCommonInfoDto userCommonInfoDto, UserInfoDto userInfoDto, Long userId) {
 
-		User existingUser = userRepository.findById(userId)
-			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 유저입니다."));
+		User existingUser = checkUser(userId);
 
-		if (userRepository.existsByNicknameAndIdNot(userCommonInfoDto.nickname(), existingUser.getId())) {
-			throw new DuplicateRequestException("이미 존재하는 닉네임입니다.");
-		}
-
-		existingUser.updateUserCommonInfo(userCommonInfoDto.nickname(), userCommonInfoDto.ImageUrl());
+		userRepository.findByNickname(userCommonInfoDto.nickname())
+			.ifPresent(user -> {
+				if(!user.getId().equals(userId)) {
+					throw new DuplicateResourceException("이미 사용중인 닉네임입니다.");
+				}
+			});
 
 		UserInfo userInfo = userInfoDto.toEntity();
+		existingUser.updateUserCommonInfo(userCommonInfoDto.nickname(), userCommonInfoDto.ImageUrl());
 		existingUser.updateUserInfo(userInfo);
 
 		return UserDto.fromEntity(userRepository.save(existingUser));
@@ -70,7 +67,6 @@ public class UserService {
 			.map(artistInfo -> artistInfo.getArtistType().getType()) // 값이 있을 때 처리
 			.orElse("User"); // 값이 없을 때 기본값
 
-		//artistType을 확인
 		return UserTypeDto.builder()
 			.userType(usertype)
 			.build();
@@ -78,21 +74,17 @@ public class UserService {
 
 	@Transactional
 	public Page<UserFollowingRes> getFollowingWithPaging(Long userId, Pageable pageable) {
-
 		return userRepository.findFollowingUsers(userId, pageable)
 			.map(UserFollowingRes::fromDto);
 	}
 
 	@Transactional
 	public void followArtist(Long userId, Long artistId) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 유저입니다."));
+		User user = checkUser(userId);
+		ArtistInfo artistInfo = getArtistInfo(artistId);
 
-		ArtistInfo artistInfo = artistInfoRepository.findByUserId(artistId)
-			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 아티스트입니다."));
-
-		if (socialRepository.existsByFollowerAndFollowing(user, artistInfo)) {
-			throw new DuplicateRequestException("이미 팔로우한 아티스트입니다.");
+		if(isFollowing(user, artistInfo)) {
+			throw new DuplicateResourceException("이미 팔로우한 아티스트입니다.");
 		}
 
 		Social social = Social.builder()
@@ -105,15 +97,45 @@ public class UserService {
 
 	@Transactional
 	public void unfollowArtist(Long userId, Long artistId) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 유저입니다."));
+		User user = checkUser(userId);
+		ArtistInfo artistInfo = getArtistInfo(artistId);
 
+		if(isFollowing(user, artistInfo)) {
+			Social social = getSocial(user, artistInfo);
+			socialRepository.delete(social);
+		}
+	}
+
+	@Transactional
+	public boolean checkSignup(String email) {
+		return userRepository.findByUserInfoPhone(email).isPresent();
+	}
+
+	private User checkUser(Long userId) {
+		return userRepository.findById(userId)
+			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 유저입니다."));
+	}
+
+	private Social getSocial(User user, ArtistInfo artistInfo) {
+		return socialRepository.findByFollowerAndFollowing(user, artistInfo)
+			.orElseThrow(() -> new ResourceNotFoundException("팔로우하지 않은 아티스트입니다."));
+	}
+
+	private boolean isFollowing(User user, ArtistInfo artistInfo) {
+		return socialRepository.existsByFollowerAndFollowing(user, artistInfo);
+	}
+
+	private ArtistInfo getArtistInfo(Long artistId) {
 		ArtistInfo artistInfo = artistInfoRepository.findByUserId(artistId)
 			.orElseThrow(() -> new ResourceNotFoundException("존재하지 않는 아티스트입니다."));
+		return artistInfo;
+	}
 
-		Social social = socialRepository.findByFollowerAndFollowing(user, artistInfo)
-			.orElseThrow(() -> new ResourceNotFoundException("팔로우하지 않은 아티스트입니다."));
+	public Optional<User> findById(Long userId) {
+		return userRepository.findById(userId);
+	}
 
-		socialRepository.delete(social);
+	public Optional<User> findByEmail(String email) {
+		return userRepository.findByUserInfoEmail(email);
 	}
 }
